main_server/requirements.txt
requests
pillow
fastapi
uvicorn
httpx
python-multipart
